Integer to Long updated and RAG implemented

+++ Appointment_Microservices/src/main/java/com/example/demo1/dto/Appointment_Dto.java
@@ -3,14 +3,14 @@ package com.example.demo1.dto;
 import java.time.LocalDateTime;
 
 public class Appointment_Dto {
-    private int doctor_id;
+    private long doctor_id;
     private LocalDateTime appointment_scheduled_time;
     private String disease_category;
     private String disease_description;
-    public int getDoctor_id() {
+    public long getDoctor_id() {
         return doctor_id;
     }
-    public void setDoctor_id(int doctor_id) {
+    public void setDoctor_id(long doctor_id) {
         this.doctor_id = doctor_id;
     }
     public LocalDateTime getAppointment_scheduled_time() {
@@ -31,7 +31,7 @@ public class Appointment_Dto {
     public void setDisease_description(String disease_description) {
         this.disease_description = disease_description;
     }
-    public Appointment_Dto(int doctor_id, LocalDateTime appointment_scheduled_time, String disease_category,
+    public Appointment_Dto(long doctor_id, LocalDateTime appointment_scheduled_time, String disease_category,
             String disease_description) {
         this.doctor_id = doctor_id;
         this.appointment_scheduled_time = appointment_scheduled_time;
